Tenth commit!

+++ app/services/auth-service/service.py
@@ -2,6 +2,7 @@ import hashlib
 import secrets
 
 import requests
+from email_validator import validate_email, EmailNotValidError
 from flask import jsonify, request, Flask, make_response
 
 import db_session
@@ -10,12 +11,62 @@ from model import AuthInfo
 app = Flask(__name__)
 
 
+def check_data(password, email, username):
+    check_flag = False
+
+    if len(password) < 8:
+        return {"status": "False", "message": "Пароль должен быть как минимум 8 символов!"}
+
+    for letter in password:
+        if letter.isdigit():
+            check_flag = True
+            break
+        if letter.isupper():
+            check_flag = True
+            break
+    if not check_flag:
+        return {"status": "False", "message": "Пароль должен содержать в себе минимум 1 заглавную букву, и одну цифру!"}
+
+    session = db_session.create_session()
+    if session.query(AuthInfo).filter(
+            AuthInfo.email == email,
+    ).first() is not None:
+        return {"status": "False", "message": "Пользователь уже существует!"}
+
+
+    if 3 > len(username):
+        return {"status": "False", "message": "Недопустимое имя пользователя! Минимальная длина имени 3 символа!"}
+
+
+    if len(username) > 15:
+        return {"status": "False", "message": "Недопустимое имя пользователя! Максимальная длина имени 15 символов!"}
+
+
+    if ' ' in username:
+        return {"status": "False",
+                "message": "Недопустимое имя пользователя! Имя пользователя не может содержать пробелы!"}
+
+
+    try:
+        validate_email(email, check_deliverability=True)
+
+
+    except EmailNotValidError:
+        return {"status": "False", "message": 'E-mail недействителен!'}
+
+    return {"status": "True", }
+
+
 @app.route('/auth/register', methods=['POST'])
 def register():
     try:
+
+        headers = dict(request.headers)
+
         session = db_session.create_session()
         req_json = request.get_json()
 
+        username = req_json['username']
         hashed_password = req_json.get('password')
         email = req_json.get('email')
         session_id = secrets.token_hex(16)
@@ -23,15 +74,27 @@ def register():
         user = AuthInfo(session_id=session_id,
                         hashed_password=hashlib.sha512(hashed_password.encode('utf-8')).hexdigest(), email=email)
         session.add(user)
+
+        if not check_data(hashed_password, email, username):
+            return make_response(
+                jsonify({"status": "False", "message": check_data(hashed_password, email, username)["message"]}))
+
         session.commit()
 
         user_id = session.query(AuthInfo).filter(AuthInfo.email == email).first()
 
-        responce = make_response(jsonify({'status': 'True', 'session_id': user.session_id, "uid": user_id.uid}), 200)
-        responce.set_cookie('session_id', user.session_id, httponly=True)
-        return responce
+        user_task = requests.post('http://user-service:5003/user/add-user', headers=headers,
+                                  json={'username': username, 'email': email, 'uid': user_id})
+
+        if user_task.json()['status'] == 'True':
+            responce = make_response(jsonify({'status': 'True', }), 200)
+            responce.set_cookie('session_id', user.session_id, httponly=True)
+            return responce
+        else:
+            return make_response(jsonify({'status': 'Something went wrong!', 'message': str(user_task.text)}))
+
     except Exception as e:
-        return make_response(jsonify({'status': 'Something went wrong!', 'message': str(e)}), 404)
+        return make_response(jsonify({'status': 'Something went wrong!', 'message': str(e)}))
 
 
 @app.route('/auth/login', methods=['POST'])
@@ -68,6 +131,10 @@ def change_password():
         id = req_json.get('uid')
         new_password = req_json.get('new_password')
 
+        if not check_data(new_password, 'example@gmail.com', 'XXXXXXXX'):
+            return make_response(jsonify(
+                {'status': 'False', 'message': check_data(new_password, 'АААААА@gmail.com', 'XXXXXXXX')["message"]}))
+
         new_hashed_password = hashlib.sha512(new_password.encode('utf-8')).hexdigest()
 
         user = session.get(AuthInfo, id)
@@ -84,17 +151,33 @@ def change_password():
 @app.route('/auth/change_email/', methods=['PATCH'])
 def change_email():
     try:
+
+        headers = dict(request.headers)
+
         session = db_session.create_session()
         req_json = request.get_json()
 
         id = req_json.get('uid')
         new_email = req_json.get('new_email')
 
-        user = session.get(AuthInfo, id)
-        user.email = new_email
-        session.commit()
+        if not check_data('example12345', new_email, 'XXXXXXXX'):
+            return make_response(jsonify(
+                {'status': 'False', 'message': check_data('example12345', new_email, 'XXXXXXXX')["message"]}))
 
-        return make_response(jsonify({'status': 'True'}), 200)
+        user_work = requests.post('http://user-service:5003/user/change-email', headers=headers, json={
+            "uid": id,
+            "new_email": new_email,
+
+        })
+
+        if user_work.json()['status'] == 'True':
+            user = session.get(AuthInfo, id)
+            user.email = new_email
+            session.commit()
+
+            return make_response(jsonify({'status': 'True'}), 200)
+        else:
+            return make_response(jsonify({"status": "False", "message": user_work.text}))
 
     except Exception as e:
 
@@ -104,14 +187,28 @@ def change_email():
 @app.route('/auth/delete_user/', methods=['DELETE'])
 def delete_user():
     try:
+
+        headers = dict(request.headers)
+
         session = db_session.create_session()
         req_json = request.get_json()
         uid = req_json.get('uid')
-        user = session.get(AuthInfo, uid)
-        session.delete(user)
-        session.commit()
 
-        return make_response(jsonify({'status': 'True'}), 200)
+        user_work = requests.post('http://user-service:5003/user/delete-user', headers=headers, json={
+
+            "uid": uid,
+        })
+
+        if user_work.json()['status'] == 'True':
+            user = session.get(AuthInfo, uid)
+            session.delete(user)
+            session.commit()
+
+            return make_response(jsonify({'status': 'True'}), 200)
+
+        else:
+
+            return make_response(jsonify({"status": "False", "message": user_work.text}))
 
     except Exception as e:
 
@@ -135,6 +232,8 @@ def validate_user():
         return make_response(jsonify({'status': 'Unknown error!', 'message': str(e)}), 401)
 
 
+'''ДОБАВИТЬ В РОУТЫ AUTH'''
+
 ROUTES = {
 
     "/chat/": "http://chat-service:5001",
@@ -149,8 +248,8 @@ ROUTES = {
 }
 
 
-@app.route("/", defaults={"path": ""}, methods=["GET", "POST", "PUT", "DELETE"])
-@app.route("/<path:path>", methods=["GET", "POST", "PUT", "DELETE"])
+@app.route("/", defaults={"path": ""}, methods=["GET", "POST", "PATCH", "DELETE"])
+@app.route("/<path:path>", methods=["GET", "POST", "PATCH", "DELETE"])
 def gateway(path):
     session_id = request.cookies.get("session_id")
